projects/web-development/JobScanner Pro/requirements.txt
requests>=2.28.0
beautifulsoup4>=4.11.0
pandas>=1.5.0
pyyaml>=6.0
rich>=12.0.0
selenium>=4.8.0
playwright>=1.30.0
python-docx>=0.8.11
PyPDF2>=3.0.0
nltk>=3.8
spacy>=3.5.0
textract>=1.6.5
python-magic>=0.4.27
pathlib>=1.0.1
urllib3>=1.26.0
lxml>=4.9.0
python-dateutil>=2.8.0
openpyxl>=3.1.0 